Added .driver and .headless settings for browsers
For each primary supported WebDriver browser, adds options for them to specify a `driver` (e.g. `"path/to/geckodriver.exe"`) or `headless: true`. There are more settings that could probably be added, but these are the ones I personally need :smile:.

I really wanted to have a generic function for all this but there was just a little bit too much variability in which methods were being called.